chore: move compatible test to tf
